+++ src/preload.js
@@ -17,6 +17,7 @@ contextBridge.exposeInMainWorld('skanfact', {
   openBackups: () => ipcRenderer.invoke('backups:open'),
   createBackup: (label) => ipcRenderer.invoke('backups:create', label),
   setTitle: (title) => ipcRenderer.send('window:title', title),
+  setDirty: (dirty) => ipcRenderer.send('window:dirty', !!dirty),
   listBackups: () => ipcRenderer.invoke('backups:list'),
   pickLogo: (title) => ipcRenderer.invoke('logo:pick', title),
   exportPdf: (html, suggestedName) => ipcRenderer.invoke('pdf:export', { html, suggestedName }),